Restructure micropython libs and adding print statements to functions in the microbit library that we will not implement in the simulator (#257)
