+++ src/common/telemetry_events.py
@@ -31,3 +31,12 @@ class TelemetryEvent(enum.Enum):
     MICROBIT_API_IMAGE_OTHER = "MICROBIT.API.IMAGE.OTHER"
     MICROBIT_API_IMAGE_STATIC = "MICROBIT.API.IMAGE.STATIC"
     MICROBIT_API_BUTTON = "MICROBIT.API.BUTTON"
+    MICROBIT_API_COMPASS = "MICROBIT.API.COMPASS"
+    MICROBIT_API_I2C = "MICROBIT.API.I2C"
+    MICROBIT_API_SPI = "MICROBIT.API.SPI"
+    MICROBIT_API_AUDIO = "MICROBIT.API.AUDIO"
+    MICROBIT_API_MUSIC = "MICROBIT.API.MUSIC"
+    MICROBIT_API_NEOPIXEL = "MICROBIT.API.NEOPIXEL"
+    MICROBIT_API_RADIO = "MICROBIT.API.RADIO"
+    MICROBIT_API_SPEECH = "MICROBIT.API.SPEECH"
+    MICROBIT_API_UTIME = "MICROBIT.API.UTIME"
