--- src/microbit/__model/producer_property.py
@@ -1,3 +1,0 @@
-class ProducerProperty(property):
-    def __get__(self, cls, owner):
-        return classmethod(self.fget).__get__(cls, owner)()
